Modify code structure : split the Executor for the runner. Writes firsts tests.
